Restructure project for unit tests and add example
