+++ README.md
@@ -1,2 +1,9 @@
 # counterpoint
 A program create the corresponding counterpoint given a cantus firmus
+
+## Tests
+Run unit tests like so:
+
+```bash
+python -m unittest discover
+```
