+++ examples/6_note.xml
